+++ src/main/java/net/itarray/automotion/internal/geometry/Group.java
@@ -3,4 +3,5 @@ package net.itarray.automotion.internal.geometry;
 public interface Group<V> {
     V plus(V addend);
     V minus(V subtrahend);
+    String toStringWithUnits(String units);
 }
